Manejo de excepciones
Se agregó ResponseStatus y Log en las excepciones

+++ src/main/java/com/example/demo/service/PersonService.java
@@ -1,7 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.exceptions.PersonAlreadyExists;
-import com.example.demo.exceptions.PersonNotExists;
+import com.example.demo.exceptions.PersonAlreadyExistsException;
+import com.example.demo.exceptions.PersonNotExistsException;
 import com.example.demo.model.Person;
 import com.example.demo.projections.PersonQuantity;
 import com.example.demo.repository.PersonRepository;
@@ -32,32 +32,51 @@ public class PersonService {
         return personRepository.findByDni(dni);
     }
 
-    public Person addPerson(Person newPerson) throws PersonAlreadyExists
+    public Person addPerson(Person newPerson) throws PersonAlreadyExistsException
     {
-        Person person = personRepository.save(newPerson);
-        return Optional.ofNullable(person).orElseThrow(()-> new PersonAlreadyExists("The person you're trying to add already exists."));
-
+        Person person = null;
+        if(!personRepository.existsByDni(newPerson.getDni()))
+        {
+            person = personRepository.save(newPerson);
+        }
+        return Optional.ofNullable(person).orElseThrow(()-> new PersonAlreadyExistsException("The person you're trying to add already exists."));
     }
 
-    public Person deletePerson(String dni) throws PersonNotExists
+    public Person deletePerson(String dni) throws PersonNotExistsException
     {
-        Person person = personRepository.delete(dni);
-        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExists("Couldn't delete, that person doesn't exists."));
+        Person person = null;
+        if(personRepository.existsByDni(dni))
+        {
+            person = personRepository.delete(dni);
+        }
+        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExistsException("Couldn't delete, that person doesn't exists."));
     }
 
-    public Person hireAsEmployee(String dni) throws PersonNotExists {
-        Person person = personRepository.hireAsEmployee(dni);
-        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExists("Couldn't hire, that person doesn't exists."));
+    public Person hireAsEmployee(String dni) throws PersonNotExistsException {
+        Person person = null;
+        if(personRepository.existsByDni(dni))
+        {
+            personRepository.hireAsEmployee(dni);
+        }
+        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExistsException("Couldn't hire, that person doesn't exists."));
     }
 
-    public Person hireAsVet(String dni) throws PersonNotExists {
-        Person person = personRepository.hirePersonAsVet(dni);
-        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExists("Couldn't hire, that person doesn't exists."));
+    public Person hireAsVet(String dni) throws PersonNotExistsException {
+        Person person = null;
+        if(personRepository.existsByDni(dni))
+        {
+            personRepository.hirePersonAsVet(dni);
+        }
+        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExistsException("Couldn't hire, that person doesn't exists."));
     }
 
-    public Person firePerson(String dni) throws PersonNotExists {
-        Person person = personRepository.firePerson(dni);
-        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExists("Couldn't fire, that person doesn't exists."));
+    public Person firePerson(String dni) throws PersonNotExistsException {
+        Person person = null;
+        if(personRepository.existsByDni(dni))
+        {
+            personRepository.firePerson(dni);
+        }
+        return Optional.ofNullable(person).orElseThrow(()-> new PersonNotExistsException("Couldn't fire, that person doesn't exists."));
     }
 
     public PersonQuantity getPersonQuantity()
